move selenium functions out of main (so easier to call as library)
