+++ ses/monitor_ses_api.py
@@ -1,0 +1,5 @@
+
+def monitor_jobs(isLiveSite=False):
+    '''Connect to API and monitor for new jobs'''
+    raise NotImplementedError('No API access information available for now')
+
